Fix issues of profileDetails view
