+++ app/components/ProfileDetails/profileDetails.css
@@ -1,16 +1,67 @@
 .component.profileDetails {
   width: 100%;
   height: auto;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  position: relative;
   background-color: var(--color-primary-light);
   border-radius: var(--border-radius-l);
   margin-bottom: var(--padding-l);
   box-shadow: var(--primary-shadow);
   box-sizing: border-box;
+
+  & .bubblesPattern {
+    position: absolute;
+    bottom: 0;
+    right: 3rem;
+    width: 292px;
+    height: 43px;
+    z-index: 0;
+
+    & img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+  }
+}
+
+.component.editProfile {
+  width: 100%;
+  box-sizing: border-box;
+  padding: var(--padding-l);
+  margin-bottom: var(--padding-l);
+
+  & fieldset {
+    padding-bottom: var(--padding-m);
+  }
+
+  & .fileInput {
+    margin-bottom: var(--padding-s);
+  }
+
+  & .descriptionInput {
+    height: var(--right-sidebar);
+  }
+
+  & button {
+    margin-bottom: var(--padding-s);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: column;
+    padding: var(--size-h3);
+  }
+
+  & .showFee {
+    font-size: var(--padding-xs);
+    line-height: 14px;
+  }
+}
+
+.viewProfile {
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  flex-wrap: nowrap;
+  position: relative;
 
   & .profileAvatar {
     position: relative;
@@ -121,25 +172,11 @@
       }
     }
   }
-
-  & .bubblesPattern {
-    position: absolute;
-    bottom: 0;
-    right: 3rem;
-    width: 292px;
-    height: 43px;
-    z-index: 0;
-
-    & img {
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-  }
 }
 
+
 @media (min-width: 480px) {
-  .component.profileDetails {
+  .component.viewProfile {
     flex-direction: row;
     flex-wrap: wrap;
 
@@ -162,7 +199,7 @@
 }
 
 @media (min-width: 900px) {
-  .component.profileDetails {
+  .component.viewProfile {
     & .profileName {
       width: 39%;
     }
@@ -170,7 +207,7 @@
 }
 
 @media (min-width: 1280px) {
-  .component.profileDetails {
+  .component.viewProfile {
     justify-content: space-around;
 
     & .profileAvatar {
