--- app/components/ProfileEdit/profileEdit.css
@@ -1,32 +1,0 @@
-.component.profileEdit {
-  width: 100%;
-  box-sizing: border-box;
-  padding: var(--padding-l);
-  margin-bottom: var(--padding-l);
-
-  & fieldset {
-    padding-bottom: var(--padding-m);
-  }
-
-  & .fileInput {
-    margin-bottom: var(--padding-s);
-  }
-
-  & .descriptionInput {
-    height: var(--right-sidebar);
-  }
-
-  & button {
-    margin-bottom: var(--padding-s);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: column;
-    padding: var(--size-h3);
-  }
-
-  & .showFee {
-    font-size: var(--padding-xs);
-    line-height: 14px;
-  }
-}
