--- app/components/ProfileEdit/types.ts
@@ -1,6 +1,0 @@
-import { Profile } from '~/configs';
-
-export type ProfileEditProps = {
-  profile: Profile
-  setShowForm: (value: boolean) => void;
-}
